MS-1493 Add priority to enqueue function.
In order to be able to have records from user clients, e.g. Cicero be indexed in real time (or a few seconds) it is necessary to introduce priority in enqueue and dequeue. 

Note: The schema is backward compatible but the rawrepo-access module is NOT. Once the rawrepo databases has been patched with schema version 22 the applications can make use of rawrepo-access version 1.9-SNAPSHOT and define their own priority.

git-svn-id: https://svn.dbc.dk/repos/rawrepo/trunk@117729 b87a6194-a221-0410-ad6f-3b8ec35bba13