Fix slow index hit
The 'sort by' column in dequeue didn't match the order to those column in the index. This caused the database to constantly sort all rows in the table. For 4 million rows on basismig db this took around 30 seconds. After changing the order to the column in the index this has been reduced to around 4 ms.

Note: The v22 schema has only been applied to the basismig environment, so we decided to sneak in the update to the existing version, instead of adding a new schema version.

git-svn-id: https://svn.dbc.dk/repos/rawrepo/trunk@117905 b87a6194-a221-0410-ad6f-3b8ec35bba13