Add record and queue overview to Hydra
Based on BDM's story MS-1387 a new page has been added to Hydra which 
shows all records by agencies.
In addition to showing the records the page also shows what is on queue.

According to the story it should be possible to sort the data, however
that has not been implemented this time around.

git-svn-id: https://svn.dbc.dk/repos/rawrepo/trunk@117916 b87a6194-a221-0410-ad6f-3b8ec35bba13