Fix error when expanding multi volume records
The old logic was that the record returned from the expand call would be the last common record in the supplied collection.
That works fine as long as there is only one common record in the collection. However with multi volume records it no longer works.

So now a recordId has to be supplied to the expand call

git-svn-id: https://svn.dbc.dk/repos/rawrepo/trunk@118102 b87a6194-a221-0410-ad6f-3b8ec35bba13