MS-1484 Remove unused providers
In preparation of removing unused providers we need to know which ones 
are used. When enqueue is called the provider name is added to a new log
table and provider count incremented.

This change is so small that we don't create a new schema, but rather
patch the databases by hand.

When we have enough data this temporary code will be removed along 
with the unused providers

git-svn-id: https://svn.dbc.dk/repos/rawrepo/trunk@118401 b87a6194-a221-0410-ad6f-3b8ec35bba13