Add records_summary and stop deleting archived records
This changelist is related to MS-1580 and MS-1588

It has several times been a problem that archived records are deleted
after a while. This new database version no longer deletes archived
records. This might mean the archive table will grow very big, but that
will just have to be dealt with if it turns out to be a problem.

In addition a new table has been added: records_summary. This table 
contains a summary of the agencies in the records table. It is
updated in realtime using insert, update and delete triggers on records
table. This table will primarily be used by Hydra.

git-svn-id: https://svn.dbc.dk/repos/rawrepo/trunk@118798 b87a6194-a221-0410-ad6f-3b8ec35bba13