Simplify solr indexer build
It seems like the current maven build, which fires up a solr server
doesn't really have any affect. The build should fail because schema.xml
is not compatible with solr version 6. The solution is to just build the
project as a normal project, but still package the solr schema in a nice
zip file.

git-svn-id: https://svn.dbc.dk/repos/rawrepo/trunk@118910 b87a6194-a221-0410-ad6f-3b8ec35bba13