Remove spaces in record ids
User of rawrepo-maintain has requested that all spaces are removed from
the record ids when using the tool.

Note: Only values entered in the textfield will be trimmed. If loading
record ids from a file should be trimmed as well then that will have to
be another change.

Note 2: The index.html file is in serious need of a beautification which
will be done in a following submit.

git-svn-id: https://svn.dbc.dk/repos/rawrepo/trunk@119245 b87a6194-a221-0410-ad6f-3b8ec35bba13