Use Date instead just the date string
The Instant object which the javascript receives has the time of GMT+0
but with time zone, which means the hour will be off by 1 or 2 hours
if the time is simply printed.
By converting the value to a Date the time zone is saved and the print
has the correct time.