Allow 870971 in agencyFor for FBS libraries
When all children for a record is put on queue it is not possible to see
if the child is a 870970 or 870971 record. This means that when a FBS
library changes their enrichment record all children of the common
record is enqueued including all the article record. And all those
records are queued as that FBS library.

If agencyFor is not capable of translating the agency of a library for
an article record to 870971 then agencyFor will throw an exception. It
has been decided that the exception is a bug, and agencyFor should
indeed be able to return 870971 for article records even if the incoming
agency is a FBS library.