Remove all dependencies to GracefulCache
GracefulCache now depends on a payara-micro jar. This means whenever the
rawrepo access dao is deployed to a full payara server then everything
will break. Also the access dao is used by stand-alone cli applications
which would then need to include a payara jar.

The solution to removing Graceful is to simply remove caching
altogether. When the caching is gone it doesn't make sense to use sub-
class for RelationHints so those classes were combined.